feat(axum): separate product route implementation from openapi spec
